pyproject.toml
[tool.poetry]
name = "model-compression"
version = "0.1.0"
description = ""
authors = [
    "konstantgr <konstantin.grotov@gmail.com>", 
    "slfdstrctd <slfdstrctd@yandex.ru>"
]
readme = "README.md"
packages = [
    {include = "all_models"}
]

[tool.poetry.dependencies]
python = "^3.10"
torch = {version = "^2.0.1", platform = "darwin"}
opencv-python = "^4.8.0.76"
pandas = "^2.1.0"
requests = "^2.31.0"
pillow = "^10.0.1"
scikit-learn = "^1.3.1"
tensorflow-model-optimization = "^0.7.5"
model-profiler = "^1.1.8"
jupyter = "^1.0.0"
notebook = "^7.0.4"
torchvision = "^0.15.2"
tqdm = "^4.66.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
